+++ corehq/warehouse/loaders/__init__.py
@@ -1,4 +1,5 @@
 from corehq.warehouse.loaders.domain import DomainStagingLoader, DomainDimLoader
+from corehq.warehouse.loaders.group import GroupStagingLoader, GroupDimLoader
 from corehq.warehouse.loaders.user import UserStagingLoader, UserDimLoader
 
 
@@ -8,5 +9,7 @@ def get_loader_by_slug(slug):
         DomainDimLoader,
         UserStagingLoader,
         UserDimLoader,
+        GroupStagingLoader,
+        GroupDimLoader,
     ]
     return {cls.slug: cls for cls in loaders}[slug]
